move prompt manager options to own file
